fix OSError
git-svn-id: http://dev.inghist.nl/svn/BiografischPortaal/bioport_repository/trunk@18128 a13ad8ac-ce12-0410-a602-c8c22c8f6e20

+++ bioport_repository/repository.py
@@ -302,7 +302,8 @@ class Repository(object):
         if source.url.endswith("tar.gz"):
             afile = bio.source_url.replace('file://', '')
             directory = os.path.dirname(afile)
-            shutil.rmtree(directory)
+            if os.path.isdir(directory):
+                shutil.rmtree(directory)
 
         return total, skipped
